fix: bugserver cases error for pivot table rowtree refactor
