+++ packages/vtable/src/layout/layout-helper.ts
@@ -1,6 +1,12 @@
-import type { Aggregation } from '../ts-types';
+import type { PivotTable } from '../PivotTable';
+import { IndicatorDimensionKeyPlaceholder } from '../tools/global';
+import { AggregationType } from '../ts-types';
+import type { PivotTableConstructorOptions, Aggregation } from '../ts-types';
 import type { ColumnData } from '../ts-types/list-table/layout-map/api';
+import type { IChartColumnIndicator } from '../ts-types/pivot-table/indicator/chart-indicator';
 import type { SimpleHeaderLayoutMap } from './simple-header-layout';
+import type { ITreeLayoutHeadNode } from './tree-helper';
+import { DimensionTree } from './tree-helper';
 
 export function checkHasAggregation(layoutMap: SimpleHeaderLayoutMap) {
   const columnObjects = layoutMap.columnObjects;
@@ -60,4 +66,115 @@ export function checkHasTreeDefine(layoutMap: SimpleHeaderLayoutMap) {
     }
   }
   return false;
+}
+
+export function parseColKeyRowKeyForPivotTable(table: PivotTable, options: PivotTableConstructorOptions) {
+  let columnDimensionTree;
+  let rowDimensionTree;
+  let isNeedResetColumnDimensionTree = false;
+  let isNeedResetRowDimensionTree = false;
+  if (options.columnTree) {
+    columnDimensionTree = new DimensionTree(
+      (table.internalProps.columnTree as ITreeLayoutHeadNode[]) ?? [],
+      table.layoutNodeId
+    );
+    if (
+      table.options.indicatorsAsCol !== false &&
+      !columnDimensionTree.dimensionKeys.contain(IndicatorDimensionKeyPlaceholder)
+    ) {
+      isNeedResetColumnDimensionTree = true;
+    }
+  }
+  if (options.rowTree) {
+    rowDimensionTree = new DimensionTree(
+      (table.internalProps.rowTree as ITreeLayoutHeadNode[]) ?? [],
+      table.layoutNodeId,
+      table.options.rowHierarchyType,
+      table.options.rowHierarchyType === 'tree' ? table.options.rowExpandLevel : undefined
+    );
+    if (
+      table.options.indicatorsAsCol === false &&
+      !rowDimensionTree.dimensionKeys.contain(IndicatorDimensionKeyPlaceholder)
+    ) {
+      isNeedResetRowDimensionTree = true;
+    }
+  }
+  const rowKeys = rowDimensionTree?.dimensionKeys?.count
+    ? rowDimensionTree.dimensionKeys.valueArr()
+    : options.rows?.reduce((keys: string[], rowObj) => {
+        if (typeof rowObj === 'string') {
+          keys.push(rowObj);
+        } else {
+          keys.push(rowObj.dimensionKey);
+        }
+        return keys;
+      }, []) ?? [];
+  const columnKeys = columnDimensionTree?.dimensionKeys?.count
+    ? columnDimensionTree.dimensionKeys.valueArr()
+    : options.columns?.reduce((keys: string[], columnObj) => {
+        if (typeof columnObj === 'string') {
+          keys.push(columnObj);
+        } else {
+          keys.push(columnObj.dimensionKey);
+        }
+        return keys;
+      }, []) ?? [];
+  const indicatorKeys =
+    options.indicators?.reduce((keys: string[], indicatorObj) => {
+      if (typeof indicatorObj === 'string') {
+        keys.push(indicatorObj);
+      } else {
+        keys.push(indicatorObj.indicatorKey);
+        if ((indicatorObj as IChartColumnIndicator).chartSpec) {
+          if (table.internalProps.dataConfig?.aggregationRules) {
+            table.internalProps.dataConfig?.aggregationRules.push({
+              field: indicatorObj.indicatorKey,
+              indicatorKey: indicatorObj.indicatorKey,
+              aggregationType: AggregationType.NONE
+            });
+          } else if (table.internalProps.dataConfig) {
+            table.internalProps.dataConfig.aggregationRules = [
+              {
+                field: indicatorObj.indicatorKey,
+                indicatorKey: indicatorObj.indicatorKey,
+                aggregationType: AggregationType.NONE
+              }
+            ];
+          } else {
+            table.internalProps.dataConfig = {
+              aggregationRules: [
+                {
+                  field: indicatorObj.indicatorKey,
+                  indicatorKey: indicatorObj.indicatorKey,
+                  aggregationType: AggregationType.NONE
+                }
+              ]
+            };
+          }
+        }
+      }
+      return keys;
+    }, []) ?? [];
+  if (options.rowHierarchyType === 'tree' && (options.extensionRows?.length ?? 0) >= 1) {
+    options.extensionRows?.forEach(extensionRow => {
+      const extension_rowKeys: string[] = [];
+      extensionRow.rows.forEach(row => {
+        if (typeof row === 'string') {
+          extension_rowKeys.push(row);
+        } else {
+          extension_rowKeys.push(row.dimensionKey);
+        }
+      });
+      rowKeys.push(...extension_rowKeys);
+    });
+  }
+  return {
+    rowKeys,
+    columnKeys,
+    indicatorKeys,
+    isNeedResetColumnDimensionTree,
+    isNeedResetRowDimensionTree,
+    columnDimensionTree,
+    rowDimensionTree
+  };
 }
